+++ agent/src/main/java/com/teamscale/jacoco/agent/convert/ConvertCommand.java
@@ -34,10 +34,10 @@ public class ConvertCommand implements ICommand {
 	/* package */ List<String> classDirectoriesOrZips = new ArrayList<>();
 
 	/**
-	 * Ant-style include patterns to apply during JaCoCo's traversal of class files.
+	 * Wildcard include patterns to apply during JaCoCo's traversal of class files.
 	 */
 	@Parameter(names = {"--filter", "-f"}, description = ""
-			+ "Ant-style include patterns to apply to all found class file locations during JaCoCo's traversal of class files."
+			+ "Wildcard include patterns to apply to all found class file locations during JaCoCo's traversal of class files."
 			+ " Note that zip contents are separated from zip files with @ and that you can filter only"
 			+ " class files, not intermediate folders/zips. Use with great care as missing class files"
 			+ " lead to broken coverage files! Turn on debug logging to see which locations are being filtered."
@@ -45,10 +45,10 @@ public class ConvertCommand implements ICommand {
 	/* package */ List<String> locationIncludeFilters = new ArrayList<>();
 
 	/**
-	 * Ant-style exclude patterns to apply during JaCoCo's traversal of class files.
+	 * Wildcard exclude patterns to apply during JaCoCo's traversal of class files.
 	 */
 	@Parameter(names = {"--exclude", "-e"}, description = ""
-			+ "Ant-style exclude patterns to apply to all found class file locations during JaCoCo's traversal of class files."
+			+ "Wildcard exclude patterns to apply to all found class file locations during JaCoCo's traversal of class files."
 			+ " Note that zip contents are separated from zip files with @ and that you can filter only"
 			+ " class files, not intermediate folders/zips. Use with great care as missing class files"
 			+ " lead to broken coverage files! Turn on debug logging to see which locations are being filtered."