+++ report-generator/src/main/java/com/teamscale/report/testwise/jacoco/cache/ProbesCache.java
@@ -7,6 +7,7 @@ import org.jacoco.core.data.ExecutionData;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Set;
+import java.util.function.Predicate;
 
 /**
  * Holds {@link ClassCoverageLookup}s for all analyzed classes.
@@ -59,14 +60,16 @@ public class ProbesCache {
 	 * Converts the given {@link ExecutionData} to {@link FileCoverageBuilder} using the cached lookups or null if the class
 	 * file of this class has not been included in the analysis or was not covered.
 	 */
-	public FileCoverageBuilder getCoverage(ExecutionData executionData) throws CoverageGenerationException {
+	public FileCoverageBuilder getCoverage(ExecutionData executionData, Predicate<String> locationIncludeFilter) throws CoverageGenerationException {
 		long classId = executionData.getId();
 		if (!containsClassId(classId)) {
-			logger.debug(
-					"Found coverage for a class " + executionData
-							.getName() + " that was not provided. Either you did not provide " +
-							"all relevant class files or you did not adjust the include/exclude filters on the agent to exclude " +
-							"coverage from irrelevant code.");
+			if (locationIncludeFilter.test(executionData.getName())) {
+				logger.warn(
+						"Found coverage for a class " + executionData
+								.getName() + " that was not provided. Either you did not provide " +
+								"all relevant class files or you did not adjust the include/exclude filters on the agent to exclude " +
+								"coverage from irrelevant code.");
+			}
 			return null;
 		}
 		if (!executionData.hasHits()) {