+++ teamscale-gradle-plugin/src/main/kotlin/com/teamscale/BuildVersion.kt
@@ -5,10 +5,10 @@ object BuildVersion {
 
     /** Extracts the plugin's version from the jar's manifest file. */
     val buildVersion: String
-        get() = BuildVersion::class.java.getPackage().implementationVersion ?: "0.3.6"
+        get() = BuildVersion::class.java.getPackage().implementationVersion ?: "0.3.7"
 
     /** Extracts the agent's version from the jar's manifest file. */
     val agentVersion: String
-        get() = BuildVersion::class.java.getPackage().specificationVersion ?: "11.2.0"
+        get() = BuildVersion::class.java.getPackage().specificationVersion ?: "11.3.0"
 
 }