+++ teamscale-gradle-plugin/src/main/kotlin/com/teamscale/config/AgentConfiguration.kt
@@ -110,7 +110,7 @@ class AgentConfiguration(val project: Project, val jacocoExtension: JacocoTaskEx
 
 class SerializableFilter(private val includes: List<String>?, private val excludes: List<String>?) : Serializable {
 
-    /** Returns a filter predicate that respects the configured include and exclude patterns. */
+    /** Returns a filter predicate that respects the configured wildcard include and exclude patterns. */
     fun getPredicate(): Predicate<String>? {
         return ClasspathWildcardIncludeFilter(
             includes?.joinToString(":") { "*$it".replace('/', '.') },