+++ CHANGELOG.md
@@ -6,6 +6,10 @@ We use [semantic versioning][semver]
   validation will be disabled from this version on
 - [fix] Agent uses higher timeouts (20s) for all HTTP connections to account for slow networks
 
+# 11.3.0
+- [breaking change] The convert tool now uses wildcard patterns for the class matching (was ant pattern before)
+- [breaking change] The agent returns the correct 204 and 400 status codes in Testwise Coverage mode
+
 # 11.2.0
 - [feature] The agent now also supports Java 11 code.